Trait cpu_local_preemption::CpuLocalField

pub unsafe trait CpuLocalField: Sized {
    const FIELD: PerCpuField;
}

This trait must be implemented for each field in per_cpu::PerCpuData.

Safety

This is marked unsafe because the implementor must guarantee that the associated FIELD constant is correctly specified.

For example, the implementation of this trait for CpuId must specify the FIELD const as PerCpuField::CpuId, but we cannot verify that here due to cyclic dependency issues.

Required Associated Constants

const FIELD: PerCpuField