Crate cpu_local_preemption

Provides CPU-local storage (per-CPU variables) and preemption management.

CPU-local variables cannot be used until after a given CPU has been initialized, i.e., its Local APIC (on x86_64) has been discovered and properly configured. Currently, the init() routine in this crate should be invoked by another init routine from the per_cpu crate.

Note that Rust offers the #[thread_local] attribute for thread-local storage (TLS), but there is no equivalent for CPU-local storage. On x86_64, TLS areas use the fs segment register for the TLS base, and this crate uses the gs segment register for the CPU-local base.

Structs

CpuLocal

A reference to a CPU-local variable.

PreemptionCount

A type wrapper around AtomicU8 that represents a CPU-local preemption count.

PreemptionGuard

A guard type that ensures preemption is disabled as long as it is held.

Enums

PerCpuField

The available CPU-local variables, i.e., fields in per_cpu::PerCpuData struct.

Traits

CpuLocalField

This trait must be implemented for each field in per_cpu::PerCpuData.

Functions

hold_preemption

Prevents preemption (preemptive task switching) from occurring until the returned guard object is dropped.

init

Initializes the CPU-local data region for this CPU.

preemption_enabled

Returns true if preemption is currently enabled on this CPU.